fix(owner-console): Add test-runner configuration for Storybook accessibility testing
- Add test-runner.cjs with axe-playwright integration for a11y checks
- Configure test-storybook:ci script using static build mode
- Scope a11y checks to critical impacts only to avoid false positives
- Update vite.config.js to conditionally disable PWA plugin in Storybook
- Update main.ts to filter PWA plugin from Storybook build

Note: owner-console test-storybook is not yet added to CI workflow due to
compatibility issues between @storybook/test-runner 0.24.1 and Storybook 8.6.14
with Vite. This requires further investigation. The test-runner works locally
but fails in CI with 'Cannot access StorybookTestRunnerError before initialization'.

Related: PR #1309
