+++ handoff/20250928/40_App/owner-console/.storybook/test-runner.cjs
@@ -1,0 +1,16 @@
+const { injectAxe, checkA11y } = require('axe-playwright');
+
+module.exports = {
+  async preVisit(page) {
+    await injectAxe(page);
+  },
+  async postVisit(page) {
+    await checkA11y(page, '#storybook-root', {
+      includedImpacts: ['critical'],
+      detailedReport: true,
+      detailedReportOptions: {
+        html: true,
+      },
+    });
+  },
+};
