--- handoff/20250928/40_App/owner-console/.storybook/test-runner.ts
@@ -1,31 +1,0 @@
-import type { TestRunnerConfig } from '@storybook/test-runner';
-import { injectAxe, checkA11y, configureAxe } from 'axe-playwright';
-
-const config: TestRunnerConfig = {
-  async preVisit(page) {
-    await injectAxe(page);
-  },
-  async postVisit(page) {
-    await configureAxe(page, {
-      rules: [
-        {
-          id: 'color-contrast',
-          enabled: true,
-        },
-        {
-          id: 'label',
-          enabled: true,
-        },
-      ],
-    });
-
-    await checkA11y(page, '#storybook-root', {
-      detailedReport: true,
-      detailedReportOptions: {
-        html: true,
-      },
-    });
-  },
-};
-
-export default config;
