+++ handoff/20250928/40_App/owner-console/package.json
@@ -21,7 +21,8 @@
     "typecheck": "tsc --noEmit",
     "storybook": "storybook dev -p 6007",
     "build-storybook": "storybook build",
-    "test-storybook": "test-storybook --ci"
+    "test-storybook": "test-storybook",
+    "test-storybook:ci": "concurrently -k -s first -n \"SB,TEST\" -c \"magenta,blue\" \"pnpm build-storybook --quiet && npx http-server storybook-static --port 6007 --silent\" \"wait-on tcp:127.0.0.1:6007 && pnpm test-storybook --url http://127.0.0.1:6007 --maxWorkers=2\""
   },
   "dependencies": {
     "@hookform/resolvers": "^5.0.1",
@@ -110,6 +111,7 @@
     "@vitest/coverage-v8": "^4.0.3",
     "axe-core": "^4.11.0",
     "axe-playwright": "^2.2.2",
+    "concurrently": "^9.2.1",
     "eslint": "^9.25.0",
     "eslint-plugin-i18next": "^6.1.3",
     "eslint-plugin-jsx-a11y": "^6.10.2",
@@ -117,6 +119,7 @@
     "eslint-plugin-react-refresh": "^0.4.19",
     "eslint-plugin-storybook": "^10.0.7",
     "globals": "^16.0.0",
+    "http-server": "^14.1.1",
     "jsdom": "^27.0.1",
     "lint-staged": "^16.2.6",
     "msw": "^2.12.2",
@@ -128,6 +131,7 @@
     "vite": "^6.3.5",
     "vite-plugin-pwa": "^1.1.0",
     "vitest": "^4.0.3",
+    "wait-on": "^9.0.3",
     "workbox-window": "^7.3.0"
   },
   "lint-staged": {
